src/requirements.txt
Pillow==8.4.0
PyQt5==5.15.6
